Added the clerk auth backend

+++ Project-ManageMent-Tool-server/dbController.js
@@ -1,10 +1,29 @@
 const Card = require("./card");
 const Project = require("./project");
 
+const ensureUserId = (req, res) => {
+  const userId = req.body.userId || req.query.userId;
+  if (!userId) {
+    res.status(400).json({
+      message: "Missing userId. Include Clerk user id in body or query.",
+    });
+    return null;
+  }
+  return userId;
+};
+
 // Project controllers
 exports.getAllProjects = async (req, res) => {
   try {
-    const projects = await Project.find();
+    const userId = req.auth.userId;
+    if (!userId) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
+    }
+    const projects = await Project.find({ userId: req.auth.userId }).sort({
+      createdAt: -1,
+    });
     res.status(200).json(projects);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -12,11 +31,19 @@ exports.getAllProjects = async (req, res) => {
 };
 
 exports.createProject = async (req, res) => {
-  const project = new Project({
-    title: req.body.title,
-    description: req.body.description,
-  });
   try {
+    const userId = req.auth.userId;
+    if (!userId) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
+    }
+
+    const project = new Project({
+      title: req.body.title,
+      description: req.body.description,
+      userId: req.auth.userId,
+    });
     const newProject = await project.save();
     res.status(201).json(newProject);
   } catch (error) {
@@ -26,17 +53,22 @@ exports.createProject = async (req, res) => {
 
 exports.updateProject = async (req, res) => {
   try {
-    const updatedProject = await Project.findByIdAndUpdate(
-      { _id: req.params.projectId },
-      req.body,
-      {
-        runValidators: true,
-        new: true,
-      }
-    );
-    if (!updatedProject) {
-      return res.status(404).json({ message: "Project not found" });
+    const userId = req.auth.userId;
+    if (!userId) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
     }
+    const project = await Project.findOne({
+      _id: req.params.projectId,
+      userId,
+    });
+    if (!project) return res.status(404).json({ message: "Project Not found" });
+
+    project.title = req.body.title ?? project.title;
+    project.description = req.body.description ?? project.description;
+
+    const updatedProject = await project.save();
     res.status(200).json(updatedProject);
   } catch (error) {
     res.status(400).json({ message: error.message });
@@ -45,21 +77,37 @@ exports.updateProject = async (req, res) => {
 
 exports.deleteProject = async (req, res) => {
   try {
-    const deletedProject = await Project.findByIdAndDelete({
+    const userId = req.auth.userId;
+    if (!userId) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
+    }
+
+    const deletedProject = await Project.findOneAndDelete({
       _id: req.params.projectId,
+      userId: req.auth.userId,
     });
     if (!deletedProject) {
-      return res.status(404).json({ message: "Card not found" });
+      return res.status(404).json({ message: "Project not found" });
     }
     res.status(200).json(deletedProject);
   } catch (error) {
-    json.status(400).json({ message: error.message });
+    res.status(400).json({ message: error.message });
   }
 };
 
 exports.getProjectById = async (req, res) => {
   try {
-    const project = await Project.findById(req.params.projectId);
+    const userId = req.auth.userId;
+    if (!userId)
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
+    const project = await Project.findOne({
+      _id: req.params.projectId,
+      userId: req.auth.userId,
+    });
     if (!project) {
       return res.status(404).json({ message: "Project not found" });
     }
@@ -72,6 +120,20 @@ exports.getProjectById = async (req, res) => {
 // Card controllers (project-specific)
 exports.getCardsByProject = async (req, res) => {
   try {
+    const userId = req.auth.userId;
+    if (!userId)
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
+    const project = await Project.findOne({
+      _id: req.params.projectId,
+      userId: req.auth.userId,
+    });
+    if (!project)
+      return res
+        .status(404)
+        .json({ message: "Project not found or not owned by user" });
+
     const cards = await Card.find({ projectId: req.params.projectId });
     res.status(200).json(cards);
   } catch (error) {
@@ -80,13 +142,30 @@ exports.getCardsByProject = async (req, res) => {
 };
 
 exports.createCard = async (req, res) => {
-  const card = new Card({
-    title: req.body.title,
-    description: req.body.description,
-    status: req.body.status,
-    projectId: req.params.projectId,
-  });
   try {
+    const userId = req.auth.userId;
+    if (!userId) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
+    }
+
+    const project = await Project.findOne({
+      _id: req.params.projectId,
+      userId: req.auth.userId,
+    });
+    if (!project)
+      return res
+        .status(404)
+        .json({ message: "Project not found or not owned by user" });
+
+    const card = new Card({
+      title: req.body.title,
+      description: req.body.description,
+      status: req.body.status,
+      projectId: req.params.projectId,
+      userId: req.auth.userId,
+    });
     const newCard = await card.save();
     res.status(201).json(newCard);
   } catch (error) {
@@ -96,17 +175,27 @@ exports.createCard = async (req, res) => {
 
 exports.updateCard = async (req, res) => {
   try {
-    const updatedCard = await Card.findOneAndUpdate(
-      { _id: req.params.cardId, projectId: req.params.projectId },
-      req.body,
-      {
-        new: true,
-        runValidators: true,
-      }
-    );
-    if (!updatedCard) {
-      return res.status(404).json({ message: "Card not found" });
+    const userId = req.auth.userId;
+    if (!userId) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
     }
+    const card = await Card.findOne({
+      _id: req.params.cardId,
+      projectId: req.params.projectId,
+      userId: req.auth.userId,
+    });
+    if (!card)
+      return res
+        .status(404)
+        .json({ message: "Card not found or not owned by user" });
+
+    card.title = req.body.title ?? card.title;
+    card.description = req.body.description ?? card.description;
+    card.status = req.body.status ?? card.status;
+
+    const updatedCard = await card.save();
     res.status(200).json(updatedCard);
   } catch (error) {
     res.status(400).json({ message: error.message });
@@ -115,9 +204,17 @@ exports.updateCard = async (req, res) => {
 
 exports.deleteCard = async (req, res) => {
   try {
+    const userId = req.auth.userId;
+    if (!userId) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized. User ID not found." });
+    }
+
     const deletedCard = await Card.findOneAndDelete({
       _id: req.params.cardId,
       projectId: req.params.projectId,
+      userId: req.auth.userId,
     });
     if (!deletedCard) {
       return res.status(404).json({ message: "Card not found" });
